requirements.txt
streamlit>=1.55.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
